+++ Content.Shared/_CP14/Magic/Components/CP14MagicSpellStorageComponent.cs
@@ -1,3 +1,4 @@
+using Content.Shared._CP14.Magic;
 using Robust.Shared.Prototypes;
 
 namespace Content.Shared._CP14.MagicStorage.Components;
@@ -5,7 +6,7 @@ namespace Content.Shared._CP14.MagicStorage.Components;
 /// <summary>
 /// A component that allows you to store spells in items
 /// </summary>
-[RegisterComponent, Access(typeof(CP14SharedMagicAttuningSystem))]
+[RegisterComponent, Access(typeof(CP14SharedMagicSystem))]
 public sealed partial class CP14MagicSpellStorageComponent : Component
 {
     /// <summary>
